requirements.txt
django==4.2.11
python-slugify